requirements.txt
Pillow==10.0.1
flet==0.10.3
pyepc==0.5.0
python-barcode0.15.1
pyinstaller==6.1.0
pandas==2.1.1